[Tuner] Fix smart benchmark timeout (#2511)
- Fixed issue from
[PR2475](https://github.com/nod-ai/shark-ai/pull/2475).
- Tuner's timeout is set for `subprocess.run()`, tuner now derives
candidate benchmark timeouts based on the measured runtime of the
baseline subprocess, instead of iree benchmarking result.
- Change timeout type from `int` to `float` in classes
- Fix time unit to keep consistency with function
`unit_to_microseconds()`

+++ sharktuner/dispatch_tuner/dispatch_tuner.py
@@ -19,8 +19,8 @@ class DispatchTuner(libtuner.TuningClient):
         super().__init__(tuner_context)
         self.compile_flags: list[str] = []
         self.benchmark_flags: list[str] = []
-        self.compile_timeout: Optional[int] = 16
-        self.benchmark_timeout: Optional[int] = None
+        self.compile_timeout: Optional[float] = 16
+        self.benchmark_timeout: Optional[float] = None
         self.auto_benchmark_timeout: bool = True
 
     @override
@@ -28,7 +28,7 @@ class DispatchTuner(libtuner.TuningClient):
         return self.compile_flags
 
     @override
-    def get_iree_compile_timeout_s(self) -> Optional[int]:
+    def get_iree_compile_timeout_s(self) -> Optional[float]:
         return self.compile_timeout
 
     @override
@@ -36,7 +36,7 @@ class DispatchTuner(libtuner.TuningClient):
         return self.benchmark_flags
 
     @override
-    def get_iree_benchmark_timeout_s(self) -> Optional[int]:
+    def get_iree_benchmark_timeout_s(self) -> Optional[float]:
         return self.benchmark_timeout
 
     @override
